% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_bar.R
\name{bar.catch.rel}
\alias{bar.catch.rel}
\title{Alias for \code{\link{catch_bar_rel}}}
\usage{
bar.catch.rel(
  data,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = NA,
  trim_labels = TRUE
)
}
\description{
Alias for \code{\link{catch_bar_rel}}
}
