% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_data_quality.R
\name{bar.quality}
\alias{bar.quality}
\title{Same as \code{\link{data_quality_bar}}}
\usage{
bar.quality(
  data,
  data_set = "NC",
  fishery_type = NA,
  catch_max = NA,
  show_percent_good_quality = TRUE
)
}
\description{
Same as \code{\link{data_quality_bar}}
}
