% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_bar.R
\name{catch_bar}
\alias{catch_bar}
\title{Plots catches by year as stacked bars. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the components of each bar}
\usage{
catch_bar(
  data,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = NA
)
}
\arguments{
\item{data}{A data frame containing nominal catches by \code{YEAR} and a given factor}

\item{fill_by}{The name of the column to be used to colorize the bar components}

\item{max_categories}{The number of maximum distinct categories (from the \code{fill_by} column) to be kept in the result. Everything else will be aggregated as 'All others'}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}

\item{num_legend_rows}{The number of rows to display in the legend}
}
\value{
the plot corresponding to the given input parameters
}
\description{
Plots catches by year as stacked bars. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the components of each bar
}
\examples{
catch_bar(NC_est(), SPECIES_CODE)                   # Bar chart with yearly estimated catches by species code
catch_bar(NC_est(), FLEET_CODE, max_categories = 5) # Bar chart with yearly estimated catches by fleet code, limited to the first five fleets in terms of total catches
catch_bar(NC_raw_std(years = c(2000:2010), species_group_codes = "TUNAS"), SPECIES_CODE, colors = color_table(unique_colors(16))) # Bar chart with yearly (2000-2010) raw catches by species code (for tuna species) with colors re-assigned
}
