% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_last_trends.R
\name{catch_last_trends_bar}
\alias{catch_last_trends_bar}
\title{Produces a summary chart of annual catch trends (in the temporal range expressed by the input data) categorized and colored according to a specific data field}
\usage{
catch_last_trends_bar(
  data,
  categorize_by = C_FISHERY_GROUP,
  colors = factorize_fishery_groups(fishery_group_colors_for(data))
)
}
\arguments{
\item{data}{the input dataset}

\item{categorize_by}{the data field to be used to categorize the annual catches}

\item{colors}{the colors to be used for each of the categories identified in the input data}
}
\value{
the summary chart of annual catch trends according to the provided inputs
}
\description{
Produces a summary chart of annual catch trends (in the temporal range expressed by the input data) categorized and colored according to a specific data field
}
