% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_last_trends.R
\name{catch_last_trends_bar_for}
\alias{catch_last_trends_bar_for}
\title{Produces a summary chart of annual catch trends (in the temporal range expressed by the input data) categorized according to a specific data field (with a
maximum number of user-provided categories) and uniformly colored with the provided outline and fill colors}
\usage{
catch_last_trends_bar_for(
  data,
  categorize_by = C_FLEET,
  outline = "darkgrey",
  fill = "darkgrey",
  max_categories = 4
)
}
\arguments{
\item{data}{the input dataset}

\item{categorize_by}{the data field to be used to categorize the annual catches}

\item{outline}{the outline color to be used for all of the categories identified in the input data}

\item{fill}{the fill color to be used for all of the categories identified in the input data}
}
\description{
Produces a summary chart of annual catch trends (in the temporal range expressed by the input data) categorized according to a specific data field (with a
maximum number of user-provided categories) and uniformly colored with the provided outline and fill colors
}
