% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_line.R
\name{catch_line_rel}
\alias{catch_line_rel}
\title{Plots the relative catches by year as relative line plots The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the components of each bar}
\usage{
catch_line_rel(
  data,
  color_by,
  max_categories = NA,
  colors = NA,
  plot_points = FALSE,
  num_legend_rows = NA,
  trim_labels = TRUE
)
}
\arguments{
\item{data}{A data frame containing nominal catches by \code{YEAR} and a given factor}

\item{color_by}{The name of the column to be used to colorize the line components}

\item{max_categories}{The number of maximum distinct categories (from the \code{fill_by} column) to be kept in the result. Everything else will be aggregated as 'All others'}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}

\item{plot_points}{A flag to force plotting solid dots at each data point}

\item{num_legend_rows}{The number of rows to display in the legend}

\item{trim_labels}{If \code{TRUE} trims all category labels to a maximum of 24 characters}
}
\value{
the plot corresponding to the given input parameters
}
\description{
Plots the relative catches by year as relative line plots The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the components of each bar
}
\examples{
catch_bar_expanded(NC_est(), SPECIES_CODE)                   # Expanded bar chart with yearly estimated catches by species code
catch_bar_expanded(NC_est(), FLEET_CODE, max_categories = 5) # Expanded bar chart with yearly estimated catches by fleet code, limited to the first five fleets in terms of total catches
catch_bar_expanded(NC_raw_std(years = c(2000:2010), species_group_codes = "TUNAS"), SPECIES_CODE, colors = color_table(unique_colors(16))) # Expanded bar chart with yearly (2000-2010) raw catches by species code (for tuna species) with colors re-assigned
}
