% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_pareto.R
\name{catch_pareto}
\alias{catch_pareto}
\title{Pareto plot of catches by two factors, one used to determine the categories on the X axis and another to determine the
sub-categories for each X category (typically FLEET_CODE and FISHERY_CODE).}
\usage{
catch_pareto(
  data,
  categorize_by,
  fill_by,
  max_categories = 5,
  max_fill_categories = NA,
  colors = NA,
  num_legend_rows = NA
)
}
\arguments{
\item{data}{A data frame containing nominal catches by \code{YEAR} and at least two given factors (category / fill)}

\item{categorize_by}{The name of the column to be used to assign labels to the bars}

\item{fill_by}{The name of the column to be used to colorize the bar components}

\item{max_categories}{The maximum number of categories to display on the X axis (an "All other" category will be added at the end if required)}

\item{max_fill_categories}{The number of maximum distinct categories (from the \code{fill_by} column) to be kept in the result. Everything else will be aggregated as 'All others'}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}

\item{num_legend_rows}{The number of rows to display in the legend}
}
\value{
the plot corresponding to the given input parameters
}
\description{
Pareto plot of catches by two factors, one used to determine the categories on the X axis and another to determine the
sub-categories for each X category (typically FLEET_CODE and FISHERY_CODE).
}
\examples{
catch_pareto(data, 2015, 2019, FLEET, FISHERY_GROUP, max_fill_categories = 9)
}
