% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_treemap.R
\name{catch_treemap}
\alias{catch_treemap}
\title{Plots nominal catches by species as a treemap chart. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the bars}
\usage{
catch_treemap(
  data,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = 2,
  show_percentages = TRUE,
  trim_labels = TRUE
)
}
\arguments{
\item{data}{A data frame containing nominal catches by \code{YEAR} and a given factor}

\item{fill_by}{The name of the column to be used to colorize the bar components}

\item{max_categories}{The number of maximum distinct categories (from the \code{fill_by} column) to be kept in the result. Everything else will be aggregated as 'All others'}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}

\item{num_legend_rows}{The number of rows to display in the legend}

\item{show_percentages}{\code{TRUE} to display percentages of each area over the total}

\item{trim_labels}{If \code{TRUE} trims all category labels to a maximum of 24 characters}
}
\value{
the plot corresponding to the given input parameters
}
\description{
Plots nominal catches by species as a treemap chart. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the bars
}
\examples{
catch_pie(data, fill_by = "SPECIES", colors = factorize_species(all_species_colors()))
}
