% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_grid_heat.R
\name{create_breaks}
\alias{create_breaks}
\title{Creates \code{num_intervals} quantiles on the provided \code{values}}
\usage{
create_breaks(values, num_intervals)
}
\arguments{
\item{values}{An array of values}

\item{num_intervals}{The number of intervals the values should be split into}
}
\value{
The calculated quantiles (to be used as breaks)
}
\description{
Creates \code{num_intervals} quantiles on the provided \code{values}
}
