% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_data_quality.R
\name{data_quality_bar}
\alias{data_quality_bar}
\title{Mixed bar + line plot of data quality by year and data set. Bars are further broken down by quality score (0 = high, 2, 4, 6, 8 = low) and
a line plot is superimposed to show the percentage of catches by year with a "good" quality score (i.e. between 0 and 2).}
\usage{
data_quality_bar(
  data,
  data_set = "NC",
  fishery_type = NA,
  catch_max = NA,
  show_percent_good_quality = TRUE
)
}
\arguments{
\item{data}{A data frame containing data quality records  by \code{YEAR}, \code{FLEET_CODE}, \code{FISHERY_CODE} and \code{SPECIES_CODE}}

\item{data_set}{One of the data sets for which data quality should be plot (one among \code{NC}, \code{CE} and \code{SF})}

\item{fishery_type}{Optional filter on the \code{FISHERY_TYPE_CODE} column (one among \code{ART} and \code{IND})}

\item{catch_max}{Optional, to set the maximum value (in 1,000 x MT) for the catch Y-axis. Useful when plotting data quality for different components (all, artisanal, industrial)
of the same data set}

\item{show_percent_good_quality}{Optional flag to turn ON or OFF the display of the red line showing the percentage of catches reported as having "good" quality. Defaults to \code{TRUE}}
}
\value{
the plot for the strata identified by the given input filters
}
\description{
Mixed bar + line plot of data quality by year and data set. Bars are further broken down by quality score (0 = high, 2, 4, 6, 8 = low) and
a line plot is superimposed to show the percentage of catches by year with a "good" quality score (i.e. between 0 and 2).
}
\examples{
data_quality_bar(data, "NC")
data_quality_bar(data, "NC", "ART")
data_quality_bar(data, "NC", "IND", 600)
}
