% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_data_quality.R
\name{data_quality_line}
\alias{data_quality_line}
\title{Multiple line plot of the percentage of "good" quality data (i.e. with a score between 0 and 2) by year and data set. Three lines are plot: one for artisanal fisheries, one for industrial fisheries and one
for all fisheries combined.}
\usage{
data_quality_line(data, data_set = "NC")
}
\arguments{
\item{data}{A data frame containing data quality records  by \code{C_YEAR}, \code{C_FLEET_CODE}, \code{C_FISHERY_CODE} and \code{C_SPECIES_CODE}}

\item{data_set}{One of the data sets for which data quality should be plot (one among \code{NC}, \code{CE} and \code{SF})}
}
\value{
the plot for the strata identified by the given input filters
}
\description{
Multiple line plot of the percentage of "good" quality data (i.e. with a score between 0 and 2) by year and data set. Three lines are plot: one for artisanal fisheries, one for industrial fisheries and one
for all fisheries combined.
}
\examples{
data_quality_line(data, "CE")
}
