% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_contour.R
\name{geo.contour}
\alias{geo.contour}
\title{Alias for \code{\link{geo_contourmap}}}
\usage{
geo.contour(
  data,
  categorize_by,
  longitude = "LON",
  latitude = "LAT",
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_IO = TRUE,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  colors = NA,
  uniform_fill = FALSE,
  contour_var = "ndensity",
  legend_title = NULL,
  trim_labels = TRUE
)
}
\description{
Alias for \code{\link{geo_contourmap}}
}
