% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_grid_contour.R
\name{geo.grid.contour}
\alias{geo.grid.contour}
\title{Alias for \code{\link{geo_grid_contourmap}}}
\usage{
geo.grid.contour(
  data,
  value,
  categorize_by,
  scale = 0.01,
  contour_var = "ndensity",
  colors = NA,
  max_categories = NA,
  uniform_fill = FALSE,
  use_centroid = TRUE,
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_IO = TRUE,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  standard_grid = grid_5x5,
  legend_title = NULL,
  trim_labels = TRUE
)
}
\description{
Alias for \code{\link{geo_grid_contourmap}}
}
