% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_grid_heat.R
\name{geo.grid.heatmap}
\alias{geo.grid.heatmap}
\title{Alias for \code{\link{geo_grid_heatmap}}}
\usage{
geo.grid.heatmap(
  data,
  value,
  yearly_average = TRUE,
  num_intervals = 5,
  unit = "t",
  breaks = NA,
  show_grid_borders = TRUE,
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_IO = TRUE,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  standard_grid = grid_5x5,
  palette = colorspace::heat_hcl,
  legend_title = NULL
)
}
\description{
Alias for \code{\link{geo_grid_heatmap}}
}
