% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_plot.R
\name{geo.plot}
\alias{geo.plot}
\title{Alias for \code{\link{geo_plot}}}
\usage{
geo.plot(
  data,
  categorize_by,
  longitude,
  latitude,
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  colors = NA,
  alpha = 0.7,
  legend_title = NULL
)
}
\description{
Alias for \code{\link{geo_plot}}
}
