% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_contour.R
\name{geo_contourmap}
\alias{geo_contourmap}
\title{Implementation of a contour plot based on a set of given georeferenced observations.}
\usage{
geo_contourmap(
  data,
  categorize_by,
  longitude = "LON",
  latitude = "LAT",
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_IO = TRUE,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  colors = NA,
  uniform_fill = FALSE,
  contour_var = "ndensity",
  legend_title = NULL
)
}
\arguments{
\item{data}{The data (shall contain a \code{FISHING_GROUND_CODE} column)}

\item{categorize_by}{The name of the column (in \code{data} that contains the categories)}

\item{longitude}{The name of the column (in \code{data}) that contains the \code{LONGITUDE} information}

\item{latitude}{The name of the column (in \code{data}) that contains the \code{LATITUDE} information}

\item{xlim}{The map horizontal limits (as a pair of longitude coordinates)}

\item{ylim}{The map vertical limits (as a pair of latitude coordinates)}

\item{show_IO}{when \code{TRUE}, the layers for F51 and F57 are plotted on the map}

\item{show_EEZs}{When \code{TRUE}, the overall EEZ layer is plotted on the map (not showing explicit distinctions among different EEZs)}

\item{show_high_seas}{When \code{TRUE}, the overall high seas layer is plotted on the map}

\item{colors}{The colors to assign to each \code{fill_by} category. If not specified, it is determined automatically from \code{fill_by} (as long as this is one of the standard categories for which a color mapping exists)}

\item{uniform_fill}{When \code{TRUE} each area will be colored uniformly, otherwise its coloring will be proportional to the density}

\item{contour_var}{The variable to use to determine the contour of the regions. One among \{ \code{"count"}, \code{"density"}, \code{"ndensity"} \}. See also \code{ggplot2::stat_density2d_filled}}

\item{legend_title}{The title to use for the legend (none if \code{NULL}). Defaults to the value provided for \code{FILL_BY}}
}
\value{
The \code{ggplot2} object representing the plot
}
\description{
Implementation of a contour plot based on a set of given georeferenced observations.
}
\examples{
geo_contourmap(ros.raw.SETS(fishery_group_codes = "LL"), FLEET_CODE, START_LON, START_LAT)
geo_contourmap(ros.raw.SETS(fishery_group_codes = "LL"), FLEET_CODE, START_LON, START_LAT, uniform_fill = TRUE)
geo_contourmap(ros.raw.SETS(fishery_group_codes = "LL"), FLEET_CODE, START_LON, START_LAT, contour_var = "count",  uniform_fill = TRUE)
}
