% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_grid_contour.R
\name{geo_grid_contourmap}
\alias{geo_grid_contourmap}
\title{Implementation of a contour plot based on a set of given georeferenced observations aggregated over regular grids.}
\usage{
geo_grid_contourmap(
  data,
  value,
  categorize_by,
  scale = 0.01,
  contour_var = "ndensity",
  colors = NA,
  max_categories = NA,
  uniform_fill = FALSE,
  use_centroid = TRUE,
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_IO = TRUE,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  standard_grid = grid_5x5,
  legend_title = NULL,
  trim_labels = TRUE
)
}
\arguments{
\item{data}{The data (shall contain a \code{C_FISHING_GROUND_CODE} column)}

\item{value}{The name of the column (in \code{data}) that contains the value identifying the number of observations in the grid}

\item{categorize_by}{The name of the column (in \code{data} that contains the categories)}

\item{scale}{A multiplicative factor to forcibly increase the number of observations}

\item{contour_var}{The variable to use to determine the contour of the regions. One among \{ \code{"count"}, \code{"density"}, \code{"ndensity"} \}. See also \code{ggplot2::stat_density2d_filled}}

\item{colors}{The colors to assign to each \code{fill_by} category. If not specified, it is determined automatically from \code{fill_by} (as long as this is one of the standard categories for which a color mapping exists)}

\item{max_categories}{The maximum number of \code{fill_by} categories to display. When specified, categories are sorted by their total value (in \code{data})) and only the first \code{max_categories} are kept. Everything else is assigned to a generic '\code{All other}' category.}

\item{uniform_fill}{When \code{TRUE} each area will be colored uniformly, otherwise its coloring will be proportional to the density}

\item{use_centroid}{When \code{TRUE}, pies are centered in the centroid location of the fishing ground they belong to. This might markedly differ from the default center of a regular grid when this contains a sensible proportion of land mass}

\item{xlim}{The map horizontal limits (as a pair of longitude coordinates)}

\item{ylim}{The map vertical limits (as a pair of latitude coordinates)}

\item{show_IO}{when \code{TRUE}, the layers for F51 and F57 are plotted on the map}

\item{show_EEZs}{When \code{TRUE}, the overall EEZ layer is plotted on the map (not showing explicit distinctions among different EEZs)}

\item{show_high_seas}{When \code{TRUE}, the overall high seas layer is plotted on the map}

\item{standard_grid}{Transforms the input data in order to only use grids of the provided \code{standard_grid} type (one among \{ \code{grid_1x1}, \code{grid_5x5}, \code{grid_10x10}, \code{grid_10x20}, \code{grid_20x20}, \code{grid_30x30} \}). The transformation is based on pre-calculated mappings (see \code{[IOTCStatistics].[dbo].[CL_FISHING_GROUND_AGGREGATIONS]}) and might lead to data loss if no mapping exists for some of the input grid codes.}

\item{legend_title}{The title to use for the legend (none if \code{NULL}). Defaults to the value provided for \code{FILL_BY}}

\item{trim_labels}{If \code{TRUE} trims all category labels to a maximum of 24 characters}
}
\value{
The \code{ggplot2} object representing the plot
}
\description{
Implementation of a contour plot based on a set of given georeferenced observations aggregated over regular grids.
}
\examples{
geo_grid_contourmap(ros.SETS(fishery_group_codes = "LL"), C_FLEET_CODE, C_START_LON, C_START_LAT)
geo_grid_contourmap(ros.SETS(fishery_group_codes = "LL"), C_FLEET_CODE, uniform_fill = TRUE)
geo_grid_contourmap(ros.SETS(fishery_group_codes = "LL"), C_FLEET_CODE, contour_var = "count",  uniform_fill = TRUE)
}
