% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_grid_pie.R
\name{geo_grid_piemap}
\alias{geo_grid_piemap}
\title{Implementation of gridded, georeferenced piemaps with ggplot2 and scatterpie libs.}
\usage{
geo_grid_piemap(
  data,
  value,
  fill_by,
  fill_by_codelist = NULL,
  yearly_average = TRUE,
  max_categories = NA,
  reference_value = NA,
  unit = "t",
  colors = NA,
  opacity = NA,
  fixed_radius = NA,
  use_centroid = TRUE,
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_IO_areas = TRUE,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  standard_grid = grid_5x5,
  legend_title = NULL,
  show_scatterpie_legend = TRUE,
  trim_labels = TRUE
)
}
\arguments{
\item{data}{The data (shall contain a \code{C_FISHING_GROUND_CODE} column)}

\item{value}{The name of the column (in \code{data}) that contains the value to plot}

\item{fill_by}{The name of the column (in \code{data} that contains the categories)}

\item{fill_by_codelist}{a \link[data.table]{data.table} object representing the codelist to be used for the \code{fill_by} column labels. Mandatory columns: SORT
CODE and NAME_EN. This parameter allows passing a codelist to avoid using \code{fill_by} with a labels columns, and to use a column representing codes only.}

\item{yearly_average}{If \code{TRUE} uses the yearly average to plot the pies (assuming that the input data contains a \code{C_YEAR} column)}

\item{max_categories}{The maximum number of \code{fill_by} categories to display. When specified, categories are sorted by their total value (in \code{data})) and only the first \code{max_categories} are kept. Everything else is assigned to a generic '\code{All other}' category.}

\item{reference_value}{The reference value for the standard pie. If not specified, it will be assumed to equal the 99\% of all values}

\item{unit}{The value unit (for display purposes)}

\item{colors}{The colors to assign to each \code{fill_by} category. If not specified, it is determined automatically from \code{fill_by} (as long as this is one of the standard categories for which a color mapping exists)}

\item{opacity}{Sets the opacity of the pies appearing in the plot (defaults to 1, i.e. fully opaque)}

\item{fixed_radius}{Sets a fixed value for the radius of the pies (in degrees). If no \code{opacity} is specified, then the \code{opacity} of the pie will depend on its relative total value}

\item{xlim}{The map horizontal limits (as a pair of longitude coordinates)}

\item{ylim}{The map vertical limits (as a pair of latitude coordinates)}

\item{show_IO_areas}{when \code{TRUE}, the layers for F51 and F57 are plotted on the map}

\item{show_EEZs}{When \code{TRUE}, the overall EEZ layer is plotted on the map (not showing explicit distinctions among different EEZs)}

\item{show_high_seas}{When \code{TRUE}, the overall high seas layer is plotted on the map}

\item{standard_grid}{Transforms the input data in order to only use grids of the provided \code{standard_grid} type (one among \{ \code{grid_1x1}, \code{grid_5x5}, \code{grid_10x10}, \code{grid_10x20}, \code{grid_20x20}, \code{grid_30x30} \}). The transformation is based on pre-calculated mappings (see \code{[IOTCStatistics].[dbo].[CL_FISHING_GROUND_AGGREGATIONS]}) and might lead to data loss if no mapping exists for some of the input grid codes.}

\item{legend_title}{The title to use for the legend (none if \code{NULL}). Defaults to the value provided for \code{VALUE}}

\item{show_scatterpie_legend}{When \code{TRUE}, shows the scatterpie legend in the bottom-left corner of the map}

\item{trim_labels}{If \code{TRUE} trims all category labels to a maximum of 24 characters}

\item{user_centroid}{When \code{TRUE}, pies are centered in the centroid location of the fishing ground they belong to. This might markedly differ from the default center of a regular grid when this contains a sensible proportion of land mass}
}
\value{
The \code{ggplot2} object representing the plot
}
\description{
Implementation of gridded, georeferenced piemaps with ggplot2 and scatterpie libs.
}
