% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_plot.R
\name{geo_plot}
\alias{geo_plot}
\title{Implementation of a geospatial plot with ggplot2}
\usage{
geo_plot(
  data,
  categorize_by,
  longitude,
  latitude,
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  colors = NA,
  alpha = 0.7,
  legend_title = NULL,
  trim_labels = TRUE
)
}
\arguments{
\item{data}{The data (shall contain a \code{C_FISHING_GROUND_CODE} column)}

\item{categorize_by}{The name of the column (in \code{data}) that contains the categories}

\item{longitude}{The name of the column (in \code{data}) that contains the \code{LONGITUDE} information}

\item{latitude}{The name of the column (in \code{data}) that contains the \code{LATITUDE} information}

\item{xlim}{The map horizontal limits (as a pair of longitude coordinates)}

\item{ylim}{The map vertical limits (as a pair of latitude coordinates)}

\item{show_EEZs}{When \code{TRUE}, the overall EEZ layer is plotted on the map (not showing explicit distinctions among different EEZs)}

\item{show_high_seas}{When \code{TRUE}, the overall high seas layer is plotted on the map}

\item{colors}{The colors to assign to each \code{fill_by} category. If not specified, it is determined automatically from \code{fill_by} (as long as this is one of the standard categories for which a color mapping exists)}

\item{alpha}{The transparency value to use when plotting the points}

\item{legend_title}{The title to use for the legend (none if \code{NULL}). Defaults to the value provided for \code{FILL_BY}}

\item{trim_labels}{If \code{TRUE} trims all category labels to a maximum of 24 characters}
}
\value{
The \code{ggplot2} object representing the plot
}
\description{
Implementation of a geospatial plot with ggplot2
}
\examples{
geo_plot(ros.raw.SETS(years = 2018), C_FISHERY_GROUP_CODE, C_START_LON, C_START_LAT)
geo_plot(ros.)
}
