% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_heatmap.R
\name{heatmap.value}
\alias{heatmap.value}
\title{Alias for \code{\link{value_heatmap}}}
\usage{
heatmap.value(
  data,
  value,
  category_x,
  category_y,
  unit = NA,
  midpoint = NA,
  relative = NA,
  color_low = "red",
  color_mid = "yellow",
  color_high = "green"
)
}
\description{
Alias for \code{\link{value_heatmap}}
}
