% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_heatmap.R
\docType{data}
\name{HEATMAP_RELATIVE_X}
\alias{HEATMAP_RELATIVE_X}
\title{String constant to force the calculation of heatmap relative values across the X-axis}
\format{
An object of class \code{character} of length 1.
}
\usage{
HEATMAP_RELATIVE_X
}
\description{
String constant to force the calculation of heatmap relative values across the X-axis
}
\keyword{datasets}
