% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_core.R
\name{initialize_all_species_colors}
\alias{initialize_all_species_colors}
\title{Initializes the species colors using the species category code as a basis}
\usage{
initialize_all_species_colors(
  connection = (iotc.core.db.connections::getDefaultDBIHandler())()
)
}
\arguments{
\item{connection}{A connection to \code{\link{IOTDB}}}
}
\description{
Initializes the species colors using the species category code as a basis
}
