% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_grid_heat.R
\name{labels_for_breaks}
\alias{labels_for_breaks}
\title{Creates break labels (in the form \code{\( VALUE_FROM - VALUE_TO \]\)}) for a given set of breaks}
\usage{
labels_for_breaks(breaks)
}
\arguments{
\item{breaks}{The breaks to create label for}
}
\value{
The labels for the provided breaks
}
\description{
Creates break labels (in the form \code{\( VALUE_FROM - VALUE_TO \]\)}) for a given set of breaks
}
