% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_line.R
\name{line.catch.rel}
\alias{line.catch.rel}
\title{Alias for \code{\link{catch_line_rel}}}
\usage{
line.catch.rel(
  data,
  color_by,
  max_categories = NA,
  colors = NA,
  plot_points = FALSE,
  num_legend_rows = NA,
  trim_labels = TRUE
)
}
\description{
Alias for \code{\link{catch_line_rel}}
}
