% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_data_quality.R
\name{line.quality.NC}
\alias{line.quality.NC}
\title{Same as \code{\link{data_quality_bar}} with \code{data_set = "NC"}}
\usage{
line.quality.NC(data)
}
\description{
Same as \code{\link{data_quality_bar}} with \code{data_set = "NC"}
}
