% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_data_quality.R
\name{line.quality}
\alias{line.quality}
\title{Same as \code{\link{data_quality_line}}}
\usage{
line.quality(data, data_set = "NC")
}
\description{
Same as \code{\link{data_quality_line}}
}
