% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_data_quality.R
\name{line.quality.SF}
\alias{line.quality.SF}
\title{Same as \code{\link{data_quality_line}} with \code{data_set = "SF"}}
\usage{
line.quality.SF(data)
}
\description{
Same as \code{\link{data_quality_line}} with \code{data_set = "SF"}
}
