% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_line.R
\name{line.value}
\alias{line.value}
\title{Alias for \code{\link{value_line}}}
\usage{
line.value(
  data,
  value,
  time = C_YEAR,
  color_by,
  color_by_codelist = NULL,
  max_categories = NA,
  colors = NA,
  plot_points = FALSE,
  num_legend_rows = NA,
  scale = 1,
  x_axis_label = NA,
  y_axis_label = "\%",
  x_breaks_every = 5,
  trim_labels = TRUE
)
}
\description{
Alias for \code{\link{value_line}}
}
