% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_lw_misc.R
\name{LW.bin.hex}
\alias{LW.bin.hex}
\title{Alias for \code{\link{lw_hexbin}}}
\usage{
LW.bin.hex(
  data,
  facet_by = SPECIES_CODE,
  swap_axis = FALSE,
  bins = 75,
  num_breaks = 5,
  colors = NA
)
}
\description{
Alias for \code{\link{lw_hexbin}}
}
