% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_lw_misc.R
\name{LW.plot}
\alias{LW.plot}
\title{Alias for \code{\link{lw_plot}}}
\usage{
LW.plot(
  data,
  categorize_by = SPECIES_CODE,
  swap_axis = FALSE,
  jitter_x = TRUE,
  jitter_y = FALSE,
  colors = NA,
  alpha = 0.7
)
}
\description{
Alias for \code{\link{lw_plot}}
}
