% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_lw_misc.R
\name{lw_hexbin}
\alias{lw_hexbin}
\title{Hexbin plot of length / value pairs}
\usage{
lw_hexbin(
  data,
  facet_by = SPECIES_CODE,
  swap_axis = FALSE,
  bins = 75,
  num_breaks = 5,
  colors = NA
)
}
\arguments{
\item{data}{The source data that shall contain a \code{LENGTH} and a \code{WEIGHT} column}

\item{facet_by}{The data column to use to produce a faceted plot. Defaults to \code{SPECIES_CODE}}

\item{swap_axis}{If \code{TRUE} weights will be mapped to the X-axis and lengths to the Y-axis}

\item{bins}{The number of bins in both vertical and horizontal directions (see also \code{\link{geom_hex}})}

\item{num_breaks}{The number of breaks to use for the scale}

\item{colors}{An alternative palette with as many colors as the \code{num_breaks}}
}
\description{
Hexbin plot of length / value pairs
}
\examples{
lw_hexbin(ros.LW(species_codes = "BET"))
lw_hexbin(ros.LW(species_codes = "BET"), facet_by = FLEET_CODE)
}
