% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_lw_misc.R
\name{lw_smooth}
\alias{lw_smooth}
\title{Computes the smoothed conditional means of the original L-W data for a given X-Y LW plot}
\usage{
lw_smooth(plot, show_ci = TRUE, level = 0.95)
}
\arguments{
\item{plot}{The original L-W plot (produced with \code{\link{lw_plot}})}

\item{show_ci}{Whether or not display the confidence interval around the computed means (defaults to \code{TRUE})}

\item{level}{The level of confidence interval to use (defaults to \code{0.95})}
}
\description{
Computes the smoothed conditional means of the original L-W data for a given X-Y LW plot
}
\examples{
lw_smooth(lw_plot(ros.LW(species_codes = "BET", length_measure_type_codes = c("EF", "FL", "TL", "PC")), categorize_by = C_LENGTH_MEASURE_TYPE_CODE, colors = color_table(unique_colors(16))))
}
