% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_pareto.R
\name{pareto.catch}
\alias{pareto.catch}
\title{Alias for \code{\link{catch_pareto}}}
\usage{
pareto.catch(
  data,
  categorize_by,
  fill_by,
  max_categories = 5,
  max_fill_categories = NA,
  colors = NA,
  num_legend_rows = NA
)
}
\description{
Alias for \code{\link{catch_pareto}}
}
