% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_pareto.R
\name{pareto.value}
\alias{pareto.value}
\title{Alias for \code{\link{value_pareto}}}
\usage{
pareto.value(
  data,
  value,
  categorize_by,
  categorize_by_codelist = NULL,
  fill_by,
  fill_by_codelist = NULL,
  max_categories = 5,
  max_fill_categories = NA,
  colors = NA,
  num_legend_rows = NA,
  scale = 1,
  y_axis_label = "Value"
)
}
\description{
Alias for \code{\link{value_pareto}}
}
