% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_pie.R
\name{pie.catch}
\alias{pie.catch}
\title{Alias for \code{\link{catch_pie}}}
\usage{
pie.catch(
  data,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = 2,
  callouts = FALSE
)
}
\description{
Alias for \code{\link{catch_pie}}
}
