% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_pie.R
\name{pie.value}
\alias{pie.value}
\title{Alias for \code{\link{value_pie}}}
\usage{
pie.value(
  data,
  value,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = 2,
  callouts = FALSE
)
}
\description{
Alias for \code{\link{value_pie}}
}
