% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_ps_fs_ls_line.R
\name{ps_fs_ls_proportions_line}
\alias{ps_fs_ls_proportions_line}
\title{Plots the proportion of FS vs. LS catches for the most common industrial purse seine fleets (EUESP, EUFRA, SYC)}
\usage{
ps_fs_ls_proportions_line(data)
}
\arguments{
\item{data}{A data table whose \code{FISHERY_CODE} column distinguishes between PS FS and PS LS}
}
\value{
The plot
}
\description{
Plots the proportion of FS vs. LS catches for the most common industrial purse seine fleets (EUESP, EUFRA, SYC)
}
