% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_core.R
\name{save_plot}
\alias{save_plot}
\title{Saves a (gg)plot with standard width and height}
\usage{
save_plot(filename, plot, width = 6.5, height = 5, background = "white")
}
\arguments{
\item{filename}{The name of the output file}

\item{plot}{The plot to save}

\item{width}{The plot width}

\item{height}{The plot height}
}
\value{
The plot itself
}
\description{
Saves a (gg)plot with standard width and height
}
