% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/iotc_base_common_plots_size_frequency_distribution.R
\name{size_distribution}
\alias{size_distribution}
\title{Plots a faceted set of size frequencies (as histogram or line plots)}
\usage{
size_distribution(
  data,
  categorize_by = FISHERY_CODE,
  measure_type = "Fork length",
  measure_unit = "cm",
  draw_line = FALSE,
  bin_size = 1,
  colors = NA
)
}
\arguments{
\item{data}{A size-frequency data set (as returned by \code{SF_raw}, \code{SF_est} and similar methods)}

\item{categorize_by}{The name of the column to be used to categorize the facets}

\item{measure_type}{The measurement type to use as X-axis label}

\item{measure_unit}{The measurement unit (to be used in the X-axis label)}

\item{draw_line}{When \code{TRUE} a line geometry (i.e. \code{geom_line}) will be used to plot the distribution}

\item{bin_size}{The bin size}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}
}
\value{
the plot
}
\description{
Plots a faceted set of size frequencies (as histogram or line plots)
}
