% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_catch_treemap.R
\name{treemap.catch}
\alias{treemap.catch}
\title{Alias for \code{\link{catch_pie}}}
\usage{
treemap.catch(
  data,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = 2,
  show_labels = TRUE,
  show_percentages = TRUE,
  trim_labels = TRUE,
  radius = 0
)
}
\description{
Alias for \code{\link{catch_pie}}
}
