% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_heatmap.R
\name{value_heatmap}
\alias{value_heatmap}
\title{Produces a heatmap of values categorized by two factors.}
\usage{
value_heatmap(
  data,
  value,
  category_x,
  category_y,
  unit = NA,
  midpoint = NA,
  relative = NA,
  color_low = "red",
  color_mid = "yellow",
  color_high = "green"
)
}
\arguments{
\item{data}{a data table}

\item{value}{the name of the column holding the value}

\item{category_x}{the name of the column holding the X-axis category}

\item{category_y}{the name of the column holding the Y-axis category}

\item{unit}{an optional unit to be displayed for the value}

\item{midpoint}{an optional midpoint for the range of possible values}

\item{relative}{if \code{NA}, the heatmap will use the absolute values, otherwise these will be normalized across the chosen axis, namely
\code{HEATMAP_RELATIVE_X} for the X-axis, \code{HEATMAP_RELATIVE_Y} for the Y-axis and \code{HEATMAP_RELATIVE_XY} for both axis}

\item{color_low}{an optional color for the lower end of the gradient (defaults to red)}

\item{color_mid}{an optional color for the middle of the gradient (defaults to yellow)}

\item{color_high}{an optional color for the higher end of the gradient (defaults to green)}
}
\description{
Produces a heatmap of values categorized by two factors.
}
