% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_pareto.R
\name{value_pareto}
\alias{value_pareto}
\title{Pareto plot of values by two factors, one used to determine the categories on the X axis and another to determine the
sub-categories for each X category (typically C_FLEET_CODE and C_FISHERY_CODE).}
\usage{
value_pareto(
  data,
  value,
  categorize_by,
  categorize_by_codelist = NULL,
  fill_by,
  fill_by_codelist = NULL,
  max_categories = 5,
  max_fill_categories = NA,
  colors = NA,
  num_legend_rows = NA,
  scale = 1,
  y_axis_label = "Value"
)
}
\arguments{
\item{data}{A data frame containing values by \code{C_YEAR} and at least two given factors (category / fill)}

\item{value}{The name of the column holding the actual values}

\item{categorize_by}{The name of the column to be used to assign labels to the bars}

\item{categorize_by_codelist}{a \link[data.table]{data.table} object representing the codelist to be used for the \code{categorize_by} column labels. Mandatory columns: SORT
CODE and NAME_EN. This parameter allows passing a codelist to avoid using \code{categorize_by} with a labels columns, and to use a column representing codes only.}

\item{fill_by}{The name of the column to be used to colorize the bar components}

\item{fill_by_codelist}{a \link[data.table]{data.table} object representing the codelist to be used for the \code{fill_by} column labels. Mandatory columns: SORT
CODE and NAME_EN. This parameter allows passing a codelist to avoid using \code{fill_by} with a labels columns, and to use a column representing codes only.}

\item{max_categories}{The maximum number of categories to display on the X axis (an "All other" category will be added at the end if required)}

\item{max_fill_categories}{The number of maximum distinct categories (from the \code{fill_by} column) to be kept in the result. Everything else will be aggregated as 'All others'}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}

\item{num_legend_rows}{The number of rows to display in the legend}

\item{scale}{A scaling factor that should be used to divide total values to normalize their extent}

\item{y_axis_label}{The label for the Y axis}
}
\value{
the plot corresponding to the given input parameters
}
\description{
Pareto plot of values by two factors, one used to determine the categories on the X axis and another to determine the
sub-categories for each X category (typically C_FLEET_CODE and C_FISHERY_CODE).
}
