% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_pie.R
\name{value_pie}
\alias{value_pie}
\title{Plots generic 'values' as a pie chart. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the bars}
\usage{
value_pie(
  data,
  value,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = 2,
  callouts = FALSE
)
}
\arguments{
\item{data}{A data frame containing values by \code{YEAR} and a given factor}

\item{value}{The name of the column holding the actual values}

\item{fill_by}{The name of the column to be used to colorize the bar components}

\item{max_categories}{The number of maximum distinct categories (from the \code{fill_by} column) to be kept in the result. Everything else will be aggregated as 'All others'}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}

\item{num_legend_rows}{The number of rows to display in the legend}

\item{callouts}{If \code{TRUE} plots informative callouts for each category}
}
\value{
the plot corresponding to the given input parameters
}
\description{
Plots generic 'values' as a pie chart. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the bars
}
