% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_bar.R
\name{bar.value.rel}
\alias{bar.value.rel}
\title{Alias for \code{\link{value_bar_rel}}}
\usage{
bar.value.rel(
  data,
  value,
  time = YEAR,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = NA,
  x_axis_label = NA,
  y_axis_label = "Values",
  x_breaks_every = 5,
  trim_labels = TRUE
)
}
\description{
Alias for \code{\link{value_bar_rel}}
}
