% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_helpers.R
\name{factorize_colors}
\alias{factorize_colors}
\title{Produces the color factors for a data frame given a \code{FILL_BY} column}
\usage{
factorize_colors(data, fill_by = SPECIES_CODE)
}
\arguments{
\item{data}{A data frame}

\item{fill_by}{The name of the column that contains the factors to use to colorize the results. Should be one among
{\code{FISHING_GROUND_CODE}, \code{FLEET_CODE},
\code{IUCN_STATUS_CODE}, \code{SPECIES_WP_CODE}, \code{SPECIES_GROUP_CODE}, \code{SPECIES_CATEGORY_CODE}, \code{SPECIES_CODE},
\code{FISHERY_TYPE_CODE}, \code{FISHERY_GROUP_CODE}, \code{FISHERY_CODE}, \code{GEAR_CODE},
\code{QUALITY_CODE}, \code{FATE_TYPE_CODE}, \code{FATE_CODE}, \code{CONDITION_TYPE_CODE}, \code{CONDITION_CODE}, \code{RAISE_CODE}}}
}
\value{
the color factors to use
}
\description{
Produces the color factors for a data frame given a \code{FILL_BY} column
}
