% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_geo_grid_pie.R
\name{geo.grid.pie}
\alias{geo.grid.pie}
\title{Alias for \code{\link{geo_grid_piemap}}}
\usage{
geo.grid.pie(
  data,
  value,
  fill_by,
  yearly_average = TRUE,
  max_categories = NA,
  reference_value = NA,
  unit = "t",
  colors = NA,
  opacity = NA,
  fixed_radius = NA,
  use_centroid = TRUE,
  xlim = IO_map_xlim,
  ylim = IO_map_ylim,
  show_IO = TRUE,
  show_EEZs = FALSE,
  show_high_seas = FALSE,
  standard_grid = grid_5x5,
  legend_title = NULL,
  show_scatterpie_legend = TRUE
)
}
\description{
Alias for \code{\link{geo_grid_piemap}}
}
