% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_line.R
\name{line.value.rel}
\alias{line.value.rel}
\title{Alias for \code{\link{value_line_rel}}}
\usage{
line.value.rel(
  data,
  value,
  time = C_YEAR,
  color_by,
  max_categories = NA,
  colors = NA,
  plot_points = FALSE,
  num_legend_rows = NA,
  x_axis_label = NA,
  y_axis_label = "Values",
  x_breaks_every = 5,
  trim_labels = TRUE
)
}
\description{
Alias for \code{\link{value_line_rel}}
}
