% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_treemap.R
\name{treemap.value}
\alias{treemap.value}
\title{Alias for \code{\link{value_treemap}}}
\usage{
treemap.value(
  data,
  value,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = 2,
  show_percentages = TRUE,
  trim_labels = TRUE
)
}
\description{
Alias for \code{\link{value_treemap}}
}
