% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_bar.R
\name{value_bar}
\alias{value_bar}
\title{Plots generic 'values' by a time-variable (generally \code{YEAR}) as stacked bars. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the components of each bar}
\usage{
value_bar(
  data,
  value,
  time = YEAR,
  fill_by,
  max_categories = NA,
  colors = NA,
  num_legend_rows = NA,
  scale = 1,
  x_axis_label = NA,
  y_axis_label = "Value",
  x_breaks_every = 5
)
}
\arguments{
\item{data}{A data frame containing values by \code{YEAR} and a given factor}

\item{value}{The name of the column holding the actual values}

\item{time}{The name of the column representing the 'time' variable}

\item{fill_by}{The name of the column to be used to colorize the bar components}

\item{max_categories}{The number of maximum distinct categories (from the \code{fill_by} column) to be kept in the result. Everything else will be aggregated as 'All others'}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}

\item{num_legend_rows}{The number of rows to display in the legend}

\item{scale}{A scaling factor that should be used to divide total values to normalize their extent}

\item{x_axis_label}{The label for the X axis}

\item{y_axis_label}{The label for the Y axis}

\item{x_breaks_every}{The size of uniform breaks on the X axis}
}
\value{
the plot corresponding to the given input parameters
}
\description{
Plots generic 'values' by a time-variable (generally \code{YEAR}) as stacked bars. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the components of each bar
}
