% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_plots_value_bar.R
\name{value_bar_rel}
\alias{value_bar_rel}
\title{Plots relative generic 'values' by a time-variable (generally \code{C_YEAR}) as stacked bars. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the components of each bar}
\usage{
value_bar_rel(
  data,
  value,
  time = C_YEAR,
  fill_by,
  fill_by_category = NULL,
  max_categories = NA,
  colors = NA,
  num_legend_rows = NA,
  x_axis_label = NA,
  y_axis_label = "\%",
  x_breaks_every = 5,
  trim_labels = TRUE
)
}
\arguments{
\item{data}{A data frame containing values by \code{C_YEAR} and a given factor}

\item{value}{The name of the column holding the actual values}

\item{time}{The name of the column representing the 'time' variable}

\item{fill_by}{The name of the column to be used to colorize the bar components}

\item{max_categories}{The number of maximum distinct categories (from the \code{fill_by} column) to be kept in the result. Everything else will be aggregated as 'All others'}

\item{colors}{A data frame containing the colors (FILL and OUTLINE) for the factors, if set to \code{NA} these will be determined by the \code{FILL_BY} parameter}

\item{num_legend_rows}{The number of rows to display in the legend}

\item{x_axis_label}{The label for the X axis}

\item{y_axis_label}{The label for the Y axis}

\item{x_breaks_every}{The size of uniform breaks on the X axis}

\item{trim_labels}{If \code{TRUE} trims all category labels to a maximum of 24 characters}

\item{fill_by_codelist}{a \link[data.table]{data.table} object representing the codelist to be used for the \code{fill_by} column labels. Mandatory columns: SORT
CODE and NAME_EN. This parameter allows passing a codelist to avoid using \code{fill_by} with a labels columns, and to use a column representing codes only.}
}
\value{
the plot corresponding to the given input parameters
}
\description{
Plots relative generic 'values' by a time-variable (generally \code{C_YEAR}) as stacked bars. The original data should be grouped by year
and a factor (e.g. species, species group, fleet etc.) that's also used to colorize the components of each bar
}
